Add ProductWrapper, split logic between context and view
