+++ Week-1/index.html
@@ -6,7 +6,7 @@
   </head>
   <body>
     <div id="root" ></div>
-    <div id="productImage"></div>
+    <div id="product-image"></div>
     <script src="./index.js"></script>
   </body>
 </html>
